frontend_app/requirements.txt
python-codicefiscale==0.8.0
qrcode==7.4.2
requests==2.31.0
streamlit==1.26.0
